w3c-reports/12092014/mac-wk2/runned-tests/html/infrastructure/urls/terminology-0/document-base-url-expected.txt


PASS The document base URL of a document containing one or more base elements with href attributes is the frozen base URL of the first base element in the document that has an href attribute, in tree order. 
PASS The fallback base URL of a document containing no base element is the document's address. 
FAIL The fallback base URL of a document whose address is about:blank is the document base URL of the creator document. assert_equals: The document base URL should be the creator document's base URL. expected "http://localhost:8800/html/infrastructure/urls/terminology-0/document-base-url.html" but got "about:blank"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/infrastructure/urls/terminology-0/document-base-url.html:37:22
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:843:37
appendChild@[native code]
http://localhost:8800/html/infrastructure/urls/terminology-0/document-base-url.html:40:32
step@http://localhost:8800/resources/testharness.js:798:30
async_test@http://localhost:8800/resources/testharness.js:90:26
global code@http://localhost:8800/html/infrastructure/urls/terminology-0/document-base-url.html:34:15)
PASS The fallback base URL of an iframe srcdoc document is the document base URL of the document's browsing context's browsing context container's document. 
